✨ Made the page a PWA
After a lot of touble I finally got parcel and workbox to play together.
building index2.html and moving it after the build isn't nice, but there is a strange bug when using index.html, where it renames the htm, js,css files to the folder name. Needs more investigating, but I already wasted like 8h on this!